# openapi.yaml
components:
  schemas:
    AddRoomMember:
      properties:
        room_id:
          title: Room Id
          type: integer
        user_id:
          title: User Id
          type: integer
      required:
      - room_id
      - user_id
      title: AddRoomMember
      type: object
...
    CreateRoom:
      properties:
        room_name:
          title: Room Name
          type: string
        user1_id:
          title: User1 Id
          type: integer
        user2_id:
          title: User2 Id
          type: integer
      required:
      - room_name
      - user1_id
      - user2_id
      title: CreateRoom
      type: object
...
    GetMessage:
      properties:
        content:
          title: Content
          type: string
        created_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Created At
        id:
          title: Id
          type: integer
        room_id:
          title: Room Id
          type: integer
        user_icon:
          anyOf:
          - type: string
          - type: 'null'
          title: User Icon
        user_id:
          title: User Id
          type: integer
      required:
      - id
      - content
      - room_id
      - user_id
      title: GetMessage
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ResponseSchema:
      properties:
        message:
          description: "API\u64CD\u4F5C\u306E\u7D50\u679C\u3092\u8AAC\u660E\u3059\u308B\
            \u30E1\u30C3\u30BB\u30FC\u30B8"
          example: "\u30E1\u30E2\u306E\u66F4\u65B0\u306B\u6210\u529F\u3057\u307E\u3057\
            \u305F"
          title: Message
          type: string
      required:
      - message
      title: ResponseSchema
      type: object
    RoomBase:
      properties:
        created_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Created At
        is_group:
          default: false
          title: Is Group
          type: boolean
        room_icon:
          title: Room Icon
          type: string
        room_id:
          title: Room Id
          type: integer
        room_name:
          title: Room Name
          type: string
      required:
      - room_id
      title: RoomBase
      type: object
    SubmitMessage:
      properties:
        content:
          title: Content
          type: string
        room_id:
          title: Room Id
          type: integer
      required:
      - content
      - room_id
      title: SubmitMessage
      type: object
...
    UpdateMessage:
      properties:
        content:
          title: Content
          type: string
        id:
          title: Id
          type: integer
      required:
      - id
      - content
      title: UpdateMessage
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: token
      type: oauth2
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /harassment/anycheck:
    post:
      operationId: harassment_check_harassment_anycheck_post
      parameters:
      - in: query
        name: post_id
        required: true
        schema:
          title: Post Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Harassment Check Harassment Anycheck Post
                type: string
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Harassment Check
      tags:
      - harassment_check
  /harassment/anycheck/text:
    post:
      operationId: harassment_check_by_text_harassment_anycheck_text_post
      parameters:
      - in: query
        name: content
        required: true
        schema:
          title: Content
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Harassment Check By Text Harassment Anycheck Text
                  Post
                type: string
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Harassment Check By Text
      tags:
      - harassment_check
...
  /talk/add_room_member:
    post:
      operationId: add_room_member_talk_add_room_member_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRoomMember'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseSchema'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Add Room Member
      tags:
      - DM
  /talk/create_room:
    post:
      operationId: create_room_talk_create_room_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoom'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseSchema'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Create Room
      tags:
      - DM
  /talk/message_delete/{message_id}:
    delete:
      operationId: delete_message_talk_message_delete__message_id__delete
      parameters:
      - in: path
        name: message_id
        required: true
        schema:
          title: Message Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseSchema'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Delete Message
      tags:
      - DM
  /talk/message_submit:
    post:
      operationId: submit_message_talk_message_submit_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitMessage'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseSchema'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Submit Message
      tags:
      - DM
  /talk/message_update:
    put:
      operationId: update_talk_message_update_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMessage'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseSchema'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Update
      tags:
      - DM
  /talk/user_information/{room_id}:
    get:
      operationId: read_room_talk_user_information__room_id__get
      parameters:
      - in: path
        name: room_id
        required: true
        schema:
          title: Room Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GetMessage'
                title: Response Read Room Talk User Information  Room Id  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Read Room
      tags:
      - DM
  /talk/user_rooms:
    get:
      operationId: get_rooms_by_user_talk_user_rooms_get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RoomBase'
                title: Response Get Rooms By User Talk User Rooms Get
                type: array
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Get Rooms By User
      tags:
      - DM
...